requirements.txt
streamlit
langgraph
langchain
langchain-core
langchain-google-genai
langchain-community
python-dotenv
duckduckgo-search
requests
